requirements.txt
boto3==1.26.137
requests==2.31.0
PyYAML==6.0.1
slack-sdk==3.21.3
python-dotenv==1.0.0
splunk-sdk==1.6.19
schedule==1.2.0
aiohttp==3.8.5
prometheus-client==0.17.1 